# contracts/openapi/researchos-v1.json
{
  "components": {
    "schemas": {
      "AssetDetailView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "current_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "logical_name": {
            "title": "Logical Name",
            "type": "string"
          },
          "source_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Key"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionView"
            },
            "title": "Versions",
            "type": "array"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "logical_name",
          "source_key",
          "current_version_id",
          "status",
          "created_by",
          "created_at",
          "versions"
        ],
        "title": "AssetDetailView",
        "type": "object"
      },
      "AssetSummaryView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "current_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "logical_name": {
            "title": "Logical Name",
            "type": "string"
          },
          "source_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Key"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "logical_name",
          "source_key",
          "current_version_id",
          "status",
          "created_by",
          "created_at"
        ],
        "title": "AssetSummaryView",
        "type": "object"
      },
      "BatchCreate": {
        "properties": {
          "actor_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Actor Id",
            "type": "string"
          }
        },
        "required": [
          "actor_id"
        ],
        "title": "BatchCreate",
        "type": "object"
      },
      "BatchFinalize": {
        "properties": {
          "actor_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Actor Id",
            "type": "string"
          }
        },
        "required": [
          "actor_id"
        ],
        "title": "BatchFinalize",
        "type": "object"
      },
      "BatchView": {
        "properties": {
          "auto_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Summary"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/IngestItemView"
            },
            "title": "Items",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "uploaded_by": {
            "title": "Uploaded By",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "uploaded_by",
          "started_at",
          "completed_at",
          "status",
          "auto_summary"
        ],
        "title": "BatchView",
        "type": "object"
      },
      "Body_upload_ingest_item_api_v1_ingest_batches__batch_id__items_post": {
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "force_new_asset": {
            "default": false,
            "title": "Force New Asset",
            "type": "boolean"
          },
          "logical_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logical Name"
          },
          "source_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Key"
          }
        },
        "required": [
          "file",
          "actor_id"
        ],
        "title": "Body_upload_ingest_item_api_v1_ingest_batches__batch_id__items_post",
        "type": "object"
      },
      "FragmentView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "excerpt_hash": {
            "title": "Excerpt Hash",
            "type": "string"
          },
          "fragment_type": {
            "title": "Fragment Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "locator_json": {
            "additionalProperties": true,
            "title": "Locator Json",
            "type": "object"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "quality": {
            "title": "Quality",
            "type": "number"
          },
          "representation_id": {
            "title": "Representation Id",
            "type": "string"
          },
          "text_or_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Or Value"
          },
          "version_id": {
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "version_id",
          "representation_id",
          "ordinal",
          "fragment_type",
          "locator_json",
          "text_or_value",
          "excerpt_hash",
          "quality",
          "created_at"
        ],
        "title": "FragmentView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestOutcomeView": {
        "properties": {
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "resolution_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Status"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          }
        },
        "required": [
          "item_id",
          "asset_id",
          "version_id",
          "resolution_status",
          "status",
          "error"
        ],
        "title": "IngestOutcomeView",
        "type": "object"
      },
      "ParseRunView": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "parser_name": {
            "title": "Parser Name",
            "type": "string"
          },
          "parser_version": {
            "title": "Parser Version",
            "type": "string"
          },
          "representation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Representation Id"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version_id": {
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "version_id",
          "parser_name",
          "parser_version",
          "status",
          "representation_id",
          "error",
          "started_at",
          "completed_at"
        ],
        "title": "ParseRunView",
        "type": "object"
      },
      "RepresentationView": {
        "properties": {
          "blob_id": {
            "title": "Blob Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "generator": {
            "title": "Generator",
            "type": "string"
          },
          "generator_version": {
            "title": "Generator Version",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata_json": {
            "additionalProperties": true,
            "title": "Metadata Json",
            "type": "object"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "representation_type": {
            "title": "Representation Type",
            "type": "string"
          },
          "version_id": {
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "version_id",
          "representation_type",
          "mime_type",
          "blob_id",
          "generator",
          "generator_version",
          "metadata_json",
          "created_at"
        ],
        "title": "RepresentationView",
        "type": "object"
      },
...
      "VersionDetailView": {
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "blob_id": {
            "title": "Blob Id",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "deterministic_metadata": {
            "additionalProperties": true,
            "title": "Deterministic Metadata",
            "type": "object"
          },
          "format_kind": {
            "title": "Format Kind",
            "type": "string"
          },
          "fragments": {
            "items": {
              "$ref": "#/components/schemas/FragmentView"
            },
            "title": "Fragments",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "parent_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Version Id"
          },
          "parse_runs": {
            "items": {
              "$ref": "#/components/schemas/ParseRunView"
            },
            "title": "Parse Runs",
            "type": "array"
          },
          "representations": {
            "items": {
              "$ref": "#/components/schemas/RepresentationView"
            },
            "title": "Representations",
            "type": "array"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "asset_id",
          "parent_version_id",
          "blob_id",
          "content_hash",
          "mime_type",
          "format_kind",
          "size_bytes",
          "source_kind",
          "source_filename",
          "deterministic_metadata",
          "created_by",
          "created_at",
          "representations",
          "fragments",
          "parse_runs"
        ],
        "title": "VersionDetailView",
        "type": "object"
      },
...
      "WorkspaceCreate": {
        "properties": {
          "actor_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Actor Id",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "actor_id"
        ],
        "title": "WorkspaceCreate",
        "type": "object"
      },
      "WorkspaceView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "created_by",
          "created_at",
          "status"
        ],
        "title": "WorkspaceView",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Deterministic M1 asset foundation",
    "title": "ResearchOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/asset-versions/{version_id}": {
      "get": {
        "operationId": "get_asset_version_api_v1_asset_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asset Version"
      }
    },
    "/api/v1/assets/{asset_id}": {
      "get": {
        "operationId": "get_asset_api_v1_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asset"
      }
    },
...
    "/api/v1/ingest-batches/{batch_id}": {
      "get": {
        "operationId": "get_ingest_batch_api_v1_ingest_batches__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ingest Batch"
      }
    },
    "/api/v1/ingest-batches/{batch_id}/finalize": {
      "post": {
        "operationId": "finalize_ingest_batch_api_v1_ingest_batches__batch_id__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchFinalize"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize Ingest Batch"
      }
    },
    "/api/v1/ingest-batches/{batch_id}/items": {
      "post": {
        "operationId": "upload_ingest_item_api_v1_ingest_batches__batch_id__items_post",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_ingest_item_api_v1_ingest_batches__batch_id__items_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestOutcomeView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Ingest Item"
      }
    },
    "/api/v1/workspaces": {
      "get": {
        "operationId": "list_workspaces_api_v1_workspaces_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceView"
                  },
                  "title": "Response List Workspaces Api V1 Workspaces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Workspaces"
      },
      "post": {
        "operationId": "create_workspace_api_v1_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workspace"
      }
    },
    "/api/v1/workspaces/{workspace_id}/assets": {
      "get": {
        "operationId": "list_assets_api_v1_workspaces__workspace_id__assets_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetSummaryView"
                  },
                  "title": "Response List Assets Api V1 Workspaces  Workspace Id  Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Assets"
      }
    },
    "/api/v1/workspaces/{workspace_id}/ingest-batches": {
      "post": {
        "operationId": "create_ingest_batch_api_v1_workspaces__workspace_id__ingest_batches_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ingest Batch"
      }
    }
  }
}